Added comment support to ContentList. Moved PostListItem to ContentList and removed PostList.
